removed a warning
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@3311 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ fem/function/localfunction/temporarylocalfunction_inline.hh
@@ -112,7 +112,7 @@ namespace Dune
         needCheckGeometry_ = discreteFunctionSpace_.multipleGeometryTypes();
       }
     }
-    assert( baseFunctionSet_.numBaseFunctions() <= dofs_.size() );
+    assert( (unsigned int)baseFunctionSet_.numBaseFunctions() <= dofs_.size() );
 
     entity_ = &entity;
     assert( baseFunctionSet_.geometryType() == entity.geometry().type() );
